add decryption; bump to version 6.0.2 (#37)
