+++ src/Evento.Repository/Evento.Repository.csproj
@@ -2,11 +2,11 @@
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>6.0.1.0</Version>
+    <Version>6.0.2.0</Version>
     <Authors>Riccardo Di Nuzzo</Authors>
     <Company>Riccardo Di Nuzzo</Company>
     <Description>This library contains an EventStoreDomainRepository based on a DomainRepository base class contained in 'Evento' library. The two libraries combined are used to build Event Sourced components interacting with EventStore database</Description>
-    <PackageReleaseNotes>add encryption capabilities in the existing DomainRepositories</PackageReleaseNotes>
+    <PackageReleaseNotes>add decryption in the existing DomainRepositories</PackageReleaseNotes>
     <Copyright>Copyright 2022</Copyright>
     <PackageLicenseUrl>https://github.com/riccardone/Evento/blob/master/LICENSE</PackageLicenseUrl>
     <PackageProjectUrl>https://github.com/riccardone/Evento.Repository</PackageProjectUrl>
